pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "haonlp"
version = "0.0.1"
dependencies = [
    "hydra-core==1.3.2",
    "mlflow==2.10.0",
    "torch==2.1.1",
    "lightning==2.1.2",
    "datasets==2.15.0",
    "transformers==4.35.2",
    "onnx==1.15.0",
    "onnxruntime-gpu==1.16.3",
    "boto3==1.28.64",
    "scikit-learn==1.4.0",
    "sentencepiece==0.1.99"
]
requires-python = ">=3.10"
authors = [
    { name = "Bryant Lee", email = "leehao90301@gmail.com" },
]
maintainers = [
    { name = "Bryant Lee", email = "leehao90301@gmail.com" },
]
description = "NLP package developed by Bryant Lee."
readme = "README.md"

[project.urls]
Repository = "https://github.com/leechehao/HaoNLP.git"

[project.scripts]
haonlp-train = "haonlp.cli.train:main"
haonlp-test = "haonlp.cli.test:main"
haonlp-predict = "haonlp.cli.predict:main"

[tool.setuptools.packages.find]
where = ["src"]
